+++ src/App.tsx
@@ -1,10 +1,10 @@
 /* eslint-disable jsx-a11y/control-has-associated-label */
 import { UserWarning } from './UserWarning';
-import { Input } from './components/Input';
 import { TaskList } from './components/TaskList';
 import { Footer } from './components/Footer';
 import { CaseOfErrorMessage } from './components/CaseOfErrorMessage';
 import { useTodo } from './provider/todoProvider';
+import { Header } from './components/Header';
 
 const USER_ID = 11433;
 
@@ -19,7 +19,7 @@ export const App: React.FC = () => {
     <div className="todoapp">
       <h1 className="todoapp__title">todos</h1>
       <div className="todoapp__content">
-        <Input />
+        <Header />
         {todos && <TaskList />}
 
         {todos.length !== 0
