--- src/components/TodoCount/index.tsx
@@ -1,1 +1,0 @@
-export * from './TodoCount';
